[generator] code refinement for `init_template`
We redefine the structure of ``templates`` dir.

The ``default`` template is defined as following.
If we invoke ``couchapp init -t default myapp``,
we will get a copy of ``templates/app`` and
``templates/vendor``.

  tempaltes/
    app/
    functions/
    vendor/

There is three types of templates -- 'app', 'function',
and 'vendor'

We can identify tempalte set via name.
e.g. we have template set named as ``mytmpl``

  templates/
    ...
    mytmpl/
      app/
      functions/
      vendor/

Ref #230
Close #232